Rename purldb-public
