+++ purldb_public/settings.py
@@ -1,19 +1,15 @@
 #
 # Copyright (c) nexB Inc. and others. All rights reserved.
-# PurlDB is a trademark of nexB Inc.
+# purldb is a trademark of nexB Inc.
 # SPDX-License-Identifier: Apache-2.0
 # See http://www.apache.org/licenses/LICENSE-2.0 for the license text.
 # See https://github.com/nexB/purldb for support or download.
 # See https://aboutcode.org for more information about nexB OSS projects.
 #
 
-import os
-import sys
+from purldb.settings import *
 
 
-def command_line():
-    '''Command line entry point.'''
-    from django.core.management import execute_from_command_line
+ROOT_URLCONF = 'purldb_public.urls'
 
-    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'purldb-public.settings')
-    execute_from_command_line(sys.argv)
+WSGI_APPLICATION = 'purldb_public.wsgi.application'
